Some Audio related code
Very hapazard code. Nothing works so far.
I'm just leacing it there so that we have a starting point.

The spectrogram route seems very limited. Either we need to find a
better singal processing liobrary or implmenet the FFT transforms and
freq-time domain analysis functions ourselves. So far the lib appraoch
doesnt look promising to me.

Worst-case scenario --- We use the current musicg lib "FingerPrint" API
for sound based search. This works perfectly in terms of retireval
accuracy and speed, but code isnt serializable at one place and this
doesnt let us save to disk. To fix this, we just have to implment the
serialble class in the source code and repackge the jar. That's very
very much easier than it sounds (not kidding. It's literaly a piece of
cake). 

TODO: Decide on which approach to take based on time constraints